Create extract-download-vuln.py
Python script to exploit websites with a "force download" script which is meant to be used to download presentations or .pdfs.

This script lets you extract a big part of the backend most times. It is programmed to find all occurences of .php, .html, .js and a few other files, and downloads them via the vulnerable "force download" script.